+++ autoprotocol/harness.py
@@ -1,7 +1,7 @@
 """
 Module containing the harness module which helps with Manifest interpretation
 
-    :copyright: 2020 by The Autoprotocol Development Team, see AUTHORS
+    :copyright: 2021 by The Autoprotocol Development Team, see AUTHORS
         for more details.
     :license: BSD, see LICENSE for more details