+++ autoprotocol/liquid_handle/liquid_handle_method.py
@@ -2,7 +2,7 @@
 
 Base class for generating complex liquid handling behavior.
 
-    :copyright: 2020 by The Autoprotocol Development Team, see AUTHORS
+    :copyright: 2021 by The Autoprotocol Development Team, see AUTHORS
         for more details.
     :license: BSD, see LICENSE for more details