+++ autoprotocol/liquid_handle/mix.py
@@ -3,7 +3,7 @@
 Base LiquidHandleMethod used by Protocol.mix to generate a series of
 movements within individual wells.
 
-    :copyright: 2020 by The Autoprotocol Development Team, see AUTHORS
+    :copyright: 2021 by The Autoprotocol Development Team, see AUTHORS
         for more details.
     :license: BSD, see LICENSE for more details
 """